refactor(automations): move detectMissingFiles to separate file
